requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyMySQL==1.1.1
psycopg2==2.9.9
drf_yasg==1.21.7
django-environ==0.11.2